+++ src/platform_impl/android/mod.rs
@@ -967,6 +967,10 @@ impl Window {
 
     pub fn set_minimized(&self, _minimized: bool) {}
 
+    pub fn is_minimized(&self) -> bool {
+        false
+    }
+
     pub fn set_maximized(&self, _maximized: bool) {}
 
     pub fn is_maximized(&self) -> bool {