+++ src/platform_impl/linux/x11/window.rs
@@ -800,6 +800,23 @@ impl UnownedWindow {
         self.xconn.primary_monitor()
     }
 
+    #[inline]
+    pub fn is_minimized(&self) -> bool {
+        let state_atom = unsafe { self.xconn.get_atom_unchecked(b"_NET_WM_STATE\0") };
+        let state = self
+            .xconn
+            .get_property(self.xwindow, state_atom, ffi::XA_ATOM);
+        let hidden_atom = unsafe { self.xconn.get_atom_unchecked(b"_NET_WM_STATE_HIDDEN\0") };
+
+        match state {
+            Ok(atoms) => {
+                let hidden = atoms.iter().any(|atom: &ffi::Atom| *atom == hidden_atom);
+                hidden
+            }
+            _ => false,
+        }
+    }
+
     fn set_minimized_inner(&self, minimized: bool) -> util::Flusher<'_> {
         unsafe {
             if minimized {