+++ src/platform_impl/macos/window.rs
@@ -856,6 +856,11 @@ impl WinitWindow {
     }
 
     #[inline]
+    pub fn is_minimized(&self) -> bool {
+        self.isMiniaturized()
+    }
+
+    #[inline]
     pub fn set_maximized(&self, maximized: bool) {
         let is_zoomed = self.is_zoomed();
         if is_zoomed == maximized {