+++ src/window.rs
@@ -822,6 +822,16 @@ impl Window {
         self.window.set_minimized(minimized);
     }
 
+    /// Gets the window's current minimized state.
+    ///
+    /// ## Platform-specific
+    ///
+    /// - **iOS / Android / Web / Wayland / Orbital:** Unsupported.
+    #[inline]
+    pub fn is_minimized(&self) -> bool {
+        self.window.is_minimized()
+    }
+
     /// Sets the window to maximized or back.
     ///
     /// ## Platform-specific